+++ .prettierignore
@@ -1,0 +1,1 @@
+projects/ngrx-auto-entity/src/lib/index.ts
