+++ projects/ngrx-auto-entity/src/index.ts
@@ -2,6 +2,22 @@
  * Public API Surface of ngrx-auto-entity
  */
 
+/*
+ * Modules
+ */
+export {
+  NgrxAutoEntityModule,
+  NgRxAutoEntityRootModuleWithEffects,
+  NgRxAutoEntityRootModuleNoEntityEffects,
+  NgRxAutoEntityRootModuleNoEffects,
+  NgRxAutoEntityFeatureModule,
+  NgRxAutoEntityModuleConfig,
+  getNgRxAutoEntityMetaReducer
+} from './lib/module';
+
+/*
+ * Common models referenced throughout Auto-Entity
+ */
 export {
   IPage,
   Page,
@@ -18,10 +34,17 @@ export { IEntityDictionary, IEntityState, EntityIdentity } from './lib/util/enti
 export { IModelState, IModelClass } from './lib/util/model-state';
 export { IEntityFacade } from './lib/util/facade';
 export { ISelectorMap } from './lib/util/selector-map';
+
+/*
+ * Builders
+ */
 export { buildFacade } from './lib/util/facade-builder';
 export { buildSelectorMap } from './lib/util/selector-map-builder';
 export { buildFeatureState, buildState } from './lib/util/state-builder';
 
+/*
+ * Action Support
+ */
 export { EntityActionTypes } from './lib/actions/action-types';
 export { IEntityInfo } from './lib/actions/entity-info';
 export { EntityAction } from './lib/actions/entity-action';
@@ -30,86 +53,66 @@ export { fromEntityActions } from './lib/actions/action-operators';
 export { ofEntityType } from './lib/actions/action-operators';
 export { ofEntityAction } from './lib/actions/action-operators';
 export { isEntityActionInstance, EntityActions } from './lib/actions/entity-actions-union';
+
+/*
+ * Actions
+ */
+export { Load, LoadFailure, LoadSuccess } from './lib/actions/load-actions';
+export { LoadMany, LoadManyFailure, LoadManySuccess } from './lib/actions/load-many-actions';
+export { LoadAll, LoadAllFailure, LoadAllSuccess } from './lib/actions/load-all-actions';
+export { LoadPage, LoadPageFailure, LoadPageSuccess } from './lib/actions/load-page-actions';
+export { LoadRange, LoadRangeFailure, LoadRangeSuccess } from './lib/actions/load-range-actions';
+
+export { CreateMany, CreateManyFailure, CreateManySuccess } from './lib/actions/create-actions';
+export { Create, CreateFailure, CreateSuccess } from './lib/actions/create-actions';
+export { UpdateMany, UpdateManyFailure, UpdateManySuccess } from './lib/actions/update-actions';
+export { Update, UpdateFailure, UpdateSuccess } from './lib/actions/update-actions';
+export { UpsertMany, UpsertManyFailure, UpsertManySuccess } from './lib/actions/upsert-actions';
+export { Upsert, UpsertFailure, UpsertSuccess } from './lib/actions/upsert-actions';
+export { ReplaceMany, ReplaceManyFailure, ReplaceManySuccess } from './lib/actions/replace-actions';
+export { Replace, ReplaceFailure, ReplaceSuccess } from './lib/actions/replace-actions';
+export { DeleteMany, DeleteManyFailure, DeleteManySuccess } from './lib/actions/delete-actions';
+export { Delete, DeleteFailure, DeleteSuccess } from './lib/actions/delete-actions';
 export {
-  Load,
-  LoadFailure,
-  LoadSuccess,
-  LoadAll,
-  LoadAllFailure,
-  LoadAllSuccess,
-  LoadMany,
-  LoadManyFailure,
-  LoadManySuccess,
-  LoadPage,
-  LoadPageFailure,
-  LoadPageSuccess,
-  LoadRange,
-  LoadRangeFailure,
-  LoadRangeSuccess,
-  CreateFailure,
-  CreateSuccess,
-  Create,
-  CreateMany,
-  CreateManyFailure,
-  CreateManySuccess,
-  Update,
-  UpdateSuccess,
-  UpdateFailure,
-  UpdateMany,
-  UpdateManySuccess,
-  UpdateManyFailure,
-  Upsert,
-  UpsertSuccess,
-  UpsertFailure,
-  UpsertMany,
-  UpsertManySuccess,
-  UpsertManyFailure,
-  Replace,
-  ReplaceFailure,
-  ReplaceSuccess,
-  ReplaceMany,
-  ReplaceManyFailure,
-  ReplaceManySuccess,
-  Delete,
-  DeleteFailure,
-  DeleteSuccess,
-  DeleteByKey,
-  DeleteByKeyFailure,
-  DeleteByKeySuccess,
-  DeleteMany,
-  DeleteManyFailure,
-  DeleteManySuccess,
   DeleteManyByKeys,
   DeleteManyByKeysFailure,
   DeleteManyByKeysSuccess,
+  DeleteByKeyFailure,
+  DeleteByKeySuccess,
+  DeleteByKey
+} from './lib/actions/delete-by-key-actions';
+
+export {
   Select,
   SelectByKey,
-  SelectMany,
-  SelectMore,
-  SelectManyByKeys,
-  SelectMoreByKeys,
   Selected,
   SelectedMany,
-  Deselect,
-  DeselectMany,
-  DeselectManyByKeys,
-  DeselectAll,
-  Deselected,
-  Edit,
-  EditByKey,
-  Edited,
-  EditedByKey,
-  Change,
-  Changed,
-  EndEdit,
-  EditEnded,
-  Clear
-} from './lib/actions/actions';
+  SelectMany,
+  SelectManyByKeys,
+  SelectMore,
+  SelectMoreByKeys
+} from './lib/actions/selection-actions';
+export { Deselected, DeselectAll, DeselectManyByKeys, DeselectMany, Deselect } from './lib/actions/deselection-actions';
+export { EditEnded, EndEdit, Changed, Change, EditedByKey, Edited, EditByKey, Edit } from './lib/actions/edit-actions';
+export { Clear } from './lib/actions/actions';
 
+/*
+ * Decorators
+ */
+export { Entity } from './lib/decorators/entity';
+export { Key } from './lib/decorators/key';
+
+/*
+ * Entity Metadata and Management
+ */
 export { ENTITY_OPTS_PROP } from './lib/decorators/entity-tokens';
-export { IEffectExcept, IEntityOptions, IEntityTransformer, Entity } from './lib/decorators/entity';
+export { IEffectExcept, IEntityOptions, IEntityTransformer } from './lib/decorators/entity';
 export { IEffectExclusions } from './lib/decorators/effect-exclusions';
 export { curd, loads, extra, all, matching, except } from './lib/decorators/effect-exclusion-utils';
+
+/*
+ * Entity Metadata Utilities
+ */
 export { makeEntity } from './lib/util/make-entity';
 export {
   nameOfEntity,
@@ -119,7 +122,6 @@ export {
   entityTransforms
 } from './lib/decorators/entity-util';
 export {
-  Key,
   getKey,
   getKeyFromModel,
   getKeyFromEntity,
@@ -129,9 +131,16 @@ export {
   checkKeyName
 } from './lib/decorators/key';
 
-export { EntityOperators } from './lib/effects/operators';
-
+/*
+ * Reducer
+ */
 export { autoEntityReducer, autoEntityMetaReducer, stateNameFromAction } from './lib/reducer/reducer';
+
+/*
+ * Entity Service
+ */
+export { NgrxAutoEntityService } from './lib/service/service';
+export { IAutoEntityService } from './lib/service/interface';
 
 export {
   IEntityRangeRef,
@@ -141,13 +150,33 @@ export {
   IEntityIdentitiesRef
 } from './lib/service/refs';
 export { IEntityWithRangeInfo, IEntityWithPageInfo, IEntityError } from './lib/service/wrapper-models';
-export { IAutoEntityService } from './lib/service/interface';
-export { NgrxAutoEntityService } from './lib/service/service';
 
+/*
+ * Transformation utilities
+ */
+export {
+  transformEntityFromServer,
+  transformEntitiesFromServer,
+  transformEntityToServer,
+  transformEntitiesToServer
+} from './lib/service/transformation';
+
+/*
+ * Operators
+ */
+export { EntityOperators } from './lib/effects/operators';
+
+/*
+ * Effect Groups
+ */
 export { EntityEffects } from './lib/effects/effects-all';
 export { LoadEffects } from './lib/effects/effects-loads';
 export { CUDEffects } from './lib/effects/effects-cud';
 export { ExtraEffects } from './lib/effects/effects-extra';
+
+/*
+ * Effects
+ */
 export {
   LoadEffect,
   LoadAllEffect,
@@ -169,20 +198,3 @@ export {
   UpsertEffect,
   UpsertManyEffect
 } from './lib/effects/effects-cud-discrete';
-
-export {
-  NgrxAutoEntityModule,
-  NgRxAutoEntityRootModuleWithEffects,
-  NgRxAutoEntityRootModuleNoEntityEffects,
-  NgRxAutoEntityRootModuleNoEffects,
-  NgRxAutoEntityFeatureModule,
-  NgRxAutoEntityModuleConfig,
-  getNgRxAutoEntityMetaReducer
-} from './lib/module';
-
-export {
-  transformEntityFromServer,
-  transformEntitiesFromServer,
-  transformEntityToServer,
-  transformEntitiesToServer
-} from './lib/service/transformation';
